chore: rename utils, types and errors files
